backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
google-generativeai
